Impl [e2e-tests] QA sprint 126 done
